+++ extra_views/__init__.py
@@ -1,5 +1,7 @@
-from extra_views.formsets import FormSetView, ModelFormSetView, InlineFormSetView
-from extra_views.advanced import CreateWithInlinesView, UpdateWithInlinesView, InlineFormSet, NamedFormsetsMixin
+from extra_views.formsets import FormSetView, ModelFormSetView, \
+    InlineFormSetView, BaseFormSetMixin, BaseInlineFormSetMixin
+from extra_views.advanced import CreateWithInlinesView, \
+    UpdateWithInlinesView, InlineFormSetFactory, NamedFormsetsMixin, InlineFormSet
 from extra_views.dates import CalendarMonthView
 from extra_views.contrib.mixins import SearchableListMixin, SortableListMixin