+++ CHANGELOG.rst
@@ -1,6 +1,23 @@
 Change History
 ==============
 
+0.12.0 (?)
+-------------------
+
+Changes:
+- Removed setting of ``BaseInlineFormSetMixin.formset_class`` and
+  ``GenericInlineFormSetMixin.formset_class`` so that ``formset`` can be set in
+  ``factory_kwargs`` instead.
+- Removed ``ModelFormSetMixin.get_context_data`` and
+  ``BaseInlineFormSetMixin.get_context_data`` as this code was duplicated from
+  Django's ``MultipleObjectMixin`` and ``SingleObjectMixin`` respectively.
+- Renamed ``BaseFormSetMixin`` to ``BaseFormSetFactory``.
+- Renamed ``BaseInlineFormSetMixin`` to ``BaseInlineFormSetFactory``.
+- Renamed ``InlineFormSet`` to ``InlineFormSetFactory``.
+- Renamed ``BaseGenericInlineFormSetMixin`` to ``BaseGenericInlineFormSetFactory``.
+- Renamed ``GenericInlineFormSet`` to ``GenericInlineFormSetFactory``.
+
+
 0.11.0 (2018-04-24)
 -------------------
 Supported Versions:
@@ -30,7 +47,7 @@ Backwards-incompatible changes
 New features:
 
 - Added SuccessMessageWithInlinesMixin (#151)
-- Allow the formset prefix to be overrideen (#154)
+- Allow the formset prefix to be overridden (#154)
 
 Bug fixes: